Revert: use 63 bins, not 64 or 65
The hamming 16-bit match maps to 0.49017, which is within reason
(between the 5th and 1st percentile outcome expected under a
double-tailed binomial distribution) so there is no underlying issue
or unexpected result from the uniformification process.
However, on 65-binning this no longer falls into the central bin.
(Note that only a 1/130-th deviation from 0.5 would be required for this
to take place.) An in-depth explanation of this artifact would be
required in the figure caption, taking away focus from the objective
of the figure to concisely explain the uniformification process.

So, stick with 63-binning and explicitly discuss error elsewhere.